pyproject.toml
[tool.poetry]
name = "rqle-ai-langchain-util"
version = "0.3.0a"
description = "Library facilitating the integration of different LLM providers in LangChain (e.g. `ollama`, `Google Gemini`, etc)."
license = "Apache-2.0"
authors = ["Quentin Reul <Quentin.H.Reul@gmail.com>"]
readme = "README.md"
homepage = "https://www.linkedin.com/company/102641077"
keywords = ["langchain", "generative ai", "llm"]

[project.urls]
"Source Code" = "https://github.com/RQledotai/rqle-ai-langchain-util"
"Bug Tracker" = "https://github.com/RQledotai/rqle-ai-langchain-util/issues"

[tool.poetry.dependencies]
boto3 = "^1.34"
google-generativeai = "^0.7"
huggingface-hub = "^0.21"
langchain = "^0.2"
langchain-core = "^0.2"
langchain-community = "^0.2"
langchain-aws = "^0.1"
langchain-google-genai = "^1"
langchain_openai = "^0.1"
llm-guard = "^0"
moviepy = "^1"
oci="^2.125"
plotly = "^5.20"
pocketsphinx = "^5"
pydub = "^0.25"
python = ">=3.11,<3.12"
python-dotenv = "^1"
pytube = "^15"
speechrecognition = "^3.10"
streamlit = "^1.32"
streamlit-tags = "^1.2"
traceloop-sdk = "^0.13"
vosk = "^0.3"

[tool.poetry.group.test.dependencies]
coverage = "*"
mock = "*"
parameterized = "*"
pytest = "^7"
pytest-mock = "*"

[tool.coverage.run]
omit = [
    "tests/*"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
